Refactor FirstrunPagerAdapter/FirstRunPage to accept imageResource too
